+++ commands/base.go
@@ -19,14 +19,26 @@ package commands
 import (
 	"bytes"
 	"context"
+	"crypto/tls"
 	"fmt"
+	"io"
+	"io/ioutil"
+	"net/http"
 	"net/url"
 	"reflect"
 	"regexp"
 	"strings"
 	"text/template"
 
+	"github.com/go-openapi/runtime"
+	httptransport "github.com/go-openapi/runtime/client"
+	inventorypb "github.com/percona/pmm/api/inventorypb/json/client"
+	managementpb "github.com/percona/pmm/api/managementpb/json/client"
+	serverpb "github.com/percona/pmm/api/serverpb/json/client"
+	"github.com/percona/pmm/utils/tlsconfig"
 	"github.com/sirupsen/logrus"
+
+	"github.com/percona/pmm-admin/agentlocal"
 )
 
 // Ctx is a shared context for all requests.
@@ -122,3 +134,91 @@ func ParseCustomLabels(labels string) (map[string]string, error) {
 	}
 	return result, nil
 }
+
+type errFromNginx string
+
+func (e errFromNginx) Error() string {
+	return "response from nginx: " + string(e)
+}
+
+func (e errFromNginx) GoString() string {
+	return fmt.Sprintf("errFromNginx(%q)", string(e))
+}
+
+// SetupClients configures local and PMM Server API clients.
+func SetupClients(ctx context.Context, serverURL string) {
+	agentlocal.SetTransport(ctx, GlobalFlags.Debug || GlobalFlags.Trace)
+
+	if serverURL == "" {
+		status, err := agentlocal.GetStatus(agentlocal.DoNotRequestNetworkInfo)
+		if err != nil {
+			if err == agentlocal.ErrNotSetUp {
+				logrus.Fatalf("Failed to get PMM Server parameters from local pmm-agent: %s.\n"+
+					"Please run `pmm-admin config` with --server-url flag.", err)
+			}
+			logrus.Fatalf("Failed to get PMM Server parameters from local pmm-agent: %s.\n"+
+				"Please use --server-url flag to specify PMM Server URL.", err)
+		}
+		GlobalFlags.ServerURL = status.ServerURL
+		GlobalFlags.ServerInsecureTLS = status.ServerInsecureTLS
+	} else {
+		var err error
+		GlobalFlags.ServerURL, err = url.Parse(serverURL)
+		if err != nil {
+			logrus.Fatalf("Invalid PMM Server URL %q: %s.", serverURL, err)
+		}
+		if GlobalFlags.ServerURL.Path == "" {
+			GlobalFlags.ServerURL.Path = "/"
+		}
+		switch GlobalFlags.ServerURL.Scheme {
+		case "http", "https":
+			// nothing
+		default:
+			logrus.Fatalf("Invalid PMM Server URL %q: scheme (https:// or http://) is missing.", serverURL)
+		}
+		if GlobalFlags.ServerURL.Host == "" {
+			logrus.Fatalf("Invalid PMM Server URL %q: host is missing.", serverURL)
+		}
+	}
+
+	// use JSON APIs over HTTP/1.1
+	transport := httptransport.New(GlobalFlags.ServerURL.Host, GlobalFlags.ServerURL.Path, []string{GlobalFlags.ServerURL.Scheme})
+	if u := GlobalFlags.ServerURL.User; u != nil {
+		password, _ := u.Password()
+		transport.DefaultAuthentication = httptransport.BasicAuth(u.Username(), password)
+	}
+	transport.SetLogger(logrus.WithField("component", "server-transport"))
+	transport.SetDebug(GlobalFlags.Debug || GlobalFlags.Trace)
+	transport.Context = ctx
+
+	// set error handlers for nginx responses if pmm-managed is down
+	errorConsumer := runtime.ConsumerFunc(func(reader io.Reader, data interface{}) error {
+		b, _ := ioutil.ReadAll(reader)
+		return errFromNginx(string(b))
+	})
+	transport.Consumers = map[string]runtime.Consumer{
+		runtime.JSONMime:    runtime.JSONConsumer(),
+		runtime.HTMLMime:    errorConsumer,
+		runtime.TextMime:    errorConsumer,
+		runtime.DefaultMime: errorConsumer,
+	}
+
+	// disable HTTP/2, set TLS config
+	httpTransport := transport.Transport.(*http.Transport)
+	httpTransport.TLSNextProto = map[string]func(string, *tls.Conn) http.RoundTripper{}
+	if GlobalFlags.ServerURL.Scheme == "https" {
+		httpTransport.TLSClientConfig = tlsconfig.Get()
+		httpTransport.TLSClientConfig.ServerName = GlobalFlags.ServerURL.Hostname()
+		httpTransport.TLSClientConfig.InsecureSkipVerify = GlobalFlags.ServerInsecureTLS
+	}
+
+	inventorypb.Default.SetTransport(transport)
+	managementpb.Default.SetTransport(transport)
+	serverpb.Default.SetTransport(transport)
+}
+
+// check interfaces
+var (
+	_ error          = errFromNginx("")
+	_ fmt.GoStringer = errFromNginx("")
+)
