+++ docker-compose.yml
@@ -1,0 +1,8 @@
+---
+version: '3'
+services:
+  pmm-admin-server:
+    image: ${PMM_SERVER_IMAGE:-percona/pmm-server:2}
+    container_name: pmm-admin-server
+    environment:
+      - PATH=/root/go/bin:$PATH
